tests/wpt/web-platform-tests/feature-policy/reporting/unoptimized-lossy-images-reporting-onload.html

<!DOCTYPE html>
<html>
  <head>
    <script src='/resources/testharness.js'></script>
    <script src='/resources/testharnessreport.js'></script>
  </head>
  <body>
    <script>
var image = new Image();
image.src = "./unoptimized-image.jpg";

var check_report_format = (reports, observer) => {
  let report = reports[0];
  assert_equals(report.type, "feature-policy-violation");
  assert_equals(report.url, document.location.href);
  assert_equals(report.body.featureId, "unoptimized-lossy-images");
  assert_equals(report.body.disposition, "enforce");
};

async_test(t => {
  new ReportingObserver(t.step_func_done(check_report_format),
                        {types: ['feature-policy-violation'], buffered: true}).observe();
}, "unoptimized-images Report Format");
    </script>
  </body>
</html>
